moved functionality of all other navigatefunctions to designators
