+++ robot_smach_states/src/robot_smach_states/util/designators/navigation/__init__.py
@@ -1,2 +1,7 @@
 from navigation import *
 from waypoint_constraints import *
+from pose_constraints import *
+from symbolic_constraints import *
+from place_constraints import *
+from observe_constraints import *
+from grasp_constraints import *
