+++ robot_smach_states/src/robot_smach_states/util/designators/navigation/navigation.py
@@ -1,19 +1,10 @@
 from __future__ import absolute_import
 
 # System
-import inspect
-import pprint
 import abc
 
-# ROS
-import rospy
-from geometry_msgs.msg import *
-
 # TU/e Robotics
-from robot_skills.util.entity import Entity
 from ..core import Designator
-from ..checks import check_resolve_type
-from cb_planner_msgs_srvs.msg import OrientationConstraint, PositionConstraint
 
 
 class NavigationConstraintsDesignator(Designator):
